Eliminate build aot --interpreter flag (#17025) (#17033)
This is required for iOS debug builds, but unused otherwise. In theory,
Android debug builds could be run in this mode, but this is historically
untested and adds unnecessary complexity to the code. If ad-hoc testing
is required, it can be patched in when necessary.

This re-lands c2f016f67395e40d92a9f80bad9523ba4e777d56 with a fix to
xcode_backend.sh to eliminate the use of --interpreter.