No shrinking for BackdropFilter's cull rect (#28174)
* No shrinking for BackdropFilter's cull rect

This will be a breaking change. Our old behavior may generate confusions
for a sample app like our added golden test: Skia will shrink the cull
rect (and thus the filtered area) to the text. The new behavior will
fill the BackdropFilter to its parent/ancestor clip. This is more
in align with our clip behaviors (no clip by default).

If this breaks your app, wrap the BackdropFilter with a ClipRect.

[wip] The golden images are not uploaded yet. I'll wait for the initial
round of review to approve the golden test before uploading them.

* Statically define the callback

* Add TODO to remind the hacking code removal

* Nit fix

* Update goldens